games/pokerth-devel: Poker game written in C++/Qt (stable branch)
PokerTH is a poker game written in C++/Qt. You can play the popular "Texas
Hold'em" poker variant against up to nine computer-opponents or play internet
games with people all over the world.

This is a FreeBSD port of the stable branch. This is not intended for submission
to the FreeBSD ports tree.